+++ src/day_two/part_one.rs
@@ -77,5 +77,5 @@ pub fn run() {
     let result = games.iter().map(|x| x.id).collect::<Vec<usize>>();
     let answer: usize = result.iter().sum();
 
-    Logger::numeric_answer(answer);
+    Logger::numeric_answer(answer as u64);
 }
